confirmation on movie delete

+++ frontend/index.js
@@ -1,9 +1,9 @@
 document.addEventListener('DOMContentLoaded', () => {
     
-    createForm()
+    // createForm()
     createListForm()
-    fetchUsers()
-    userDelete()
+    // fetchUsers()
+    // userDelete()
     fetchLists()
     
     
@@ -13,75 +13,75 @@ document.addEventListener('DOMContentLoaded', () => {
 // read - fetch users index 
 const BASE_URL = "http://localhost:3000"
 
-function fetchUsers(){
-    fetch(`${BASE_URL}/users`)
-    .then(resp => resp.json())
-    .then(userData => {
-        for (const user of userData){
-            let u = new User(user.id, user.name, user.username, user.email)
-            u.renderUser();
-        }
-        // do something with the data we fetch
-    })
-}
+// function fetchUsers(){
+//     fetch(`${BASE_URL}/users`)
+//     .then(resp => resp.json())
+//     .then(userData => {
+//         for (const user of userData){
+//             let u = new User(user.id, user.name, user.username, user.email)
+//             u.renderUser();
+//         }
+//         // do something with the data we fetch
+//     })
+// }
 // create - create a new user 
-function createForm(){
-    let usersForm = document.getElementById('users-form')
+// function createForm(){
+//     let usersForm = document.getElementById('users-form')
 
-    usersForm.innerHTML += 
-    `
-    <form>
-    <h2> User Form </h2>
-    <label for="name">Name: </label>
-    <input type="text" id="name"><br>
-    <label for="username">Username: </label>
-    <input type="text" id="username"><br>
-    <label for="email">Email:</label>
-    <input type="text" id="email"><br>
-    <input type="submit" value="Create User">
-    </form>
-    `
-    usersForm.addEventListener('submit', userFormSubmission)
-}
+//     usersForm.innerHTML += 
+//     `
+//     <form>
+//     <h2> User Form </h2>
+//     <label for="name">Name: </label>
+//     <input type="text" id="name"><br>
+//     <label for="username">Username: </label>
+//     <input type="text" id="username"><br>
+//     <label for="email">Email:</label>
+//     <input type="text" id="email"><br>
+//     <input type="submit" value="Create User">
+//     </form>
+//     `
+//     usersForm.addEventListener('submit', userFormSubmission)
+// }
 
-function userFormSubmission(e){
-    e.preventDefault()
-    let name = document.getElementById('name').value
-    let username = document.getElementById('username').value
-    let email = document.getElementById('email').value
-    let user = {
-        name: name,
-        username: username,
-        email: email
-    }
-    fetch(`${BASE_URL}/users`,{
-    method: 'POST',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json'
-    },
-    body: JSON.stringify(user)
-}) 
-.then(response => response.json() )
-.then(user => {
-    let u = new User(user.id, user.name, user.username, user.email)
-    u.renderUser()
-})
-}
+// function userFormSubmission(e){
+//     e.preventDefault()
+//     let name = document.getElementById('name').value
+//     let username = document.getElementById('username').value
+//     let email = document.getElementById('email').value
+//     let user = {
+//         name: name,
+//         username: username,
+//         email: email
+//     }
+//     fetch(`${BASE_URL}/users`,{
+//     method: 'POST',
+//     headers: {
+//       'Accept': 'application/json',
+//       'Content-Type': 'application/json'
+//     },
+//     body: JSON.stringify(user)
+// }) 
+// .then(response => response.json() )
+// .then(user => {
+//     let u = new User(user.id, user.name, user.username, user.email)
+//     u.renderUser()
+// })
+// }
 
-function userDelete(){
-   userDiv = document.getElementById('users-container')
-    userDiv.addEventListener('click',(e)=>{
-        if (e.target.className === "delete-btn"){
-          fetch(`${BASE_URL}/users/${e.target.dataset.id}`, {
-            method:'DELETE'
-          })
-          .then(response => {
-            e.target.parentElement.remove()
-          })
-        }
-      })
-}
+// function userDelete(){
+//    userDiv = document.getElementById('users-container')
+//     userDiv.addEventListener('click',(e)=>{
+//         if (e.target.className === "delete-btn"){
+//           fetch(`${BASE_URL}/users/${e.target.dataset.id}`, {
+//             method:'DELETE'
+//           })
+//           .then(response => {
+//             e.target.parentElement.remove()
+//           })
+//         }
+//       })
+// }
 
 function fetchLists(){
     fetch(`${BASE_URL}/lists`)
@@ -217,24 +217,20 @@ const createMovie = (e) => {
 
   const deleteMovie = (e) => {
     e.preventDefault()
-    console.log(e.target.parentElement.parentElement)
-    
-    const configObj = {
-        method:"DELETE",
-        headers: {
-            "Content-Type": "application/json",
-            "Accept": "application/json"
-        }
+    let result = confirm("Want to delete?");
+        if (result) {
+            const configObj = {
+                method:"DELETE",
+                headers: {
+                    "Content-Type": "application/json",
+                    "Accept": "application/json"
+                }
         
+            }
+            fetch(`${BASE_URL}/movies/${e.target.dataset.movieId}`, configObj)
+                e.target.parentElement.remove()
+    
     }
-    fetch(`${BASE_URL}/movies/${e.target.dataset.movieId}`, configObj)
-        e.target.parentElement.remove()
-        
-        
-       
-    
-    
-    
     
  }
